Refs #16: WIP Refactor projects
