+++ client/build.gradle
@@ -1,0 +1,9 @@
+plugins {
+    id 'org.jetbrains.kotlin.jvm' version '1.4.20'
+    id 'java'
+}
+
+dependencies {
+    compile project(':servers:game')
+}
+
