--- servers/client/src/main/kotlin/fr/rob/client/web/handler/HomePageHandler.kt
@@ -1,9 +1,0 @@
-package fr.rob.client.web.handler
-
-import fr.rob.client.web.network.Request
-import fr.rob.client.web.network.RouteHandler
-
-class HomePageHandler : RouteHandler() {
-
-    override fun handle(request: Request): String = "Bien le bonjour"
-}
