+++ servers/web/build.gradle
@@ -5,7 +5,7 @@ plugins {
 }
 
 dependencies {
-    compile project(':servers:game')
+    compile project(':client')
 }
 
-mainClassName = 'fr.rob.client.web.Main'
+mainClassName = 'fr.rob.web.Main'
