+++ servers/web/src/main/kotlin/fr/rob/web/handler/OpcodeHandler.kt
@@ -1,11 +1,11 @@
-package fr.rob.client.web.handler
+package fr.rob.web.handler
 
 import com.google.protobuf.util.JsonFormat
-import fr.rob.client.client.network.Client
-import fr.rob.client.web.network.Request
-import fr.rob.client.web.network.RouteHandler
+import fr.rob.client.network.Client
 import fr.rob.game.domain.network.packet.Packet
 import fr.rob.game.domain.opcode.ProtobufOpcodeFunction
+import fr.rob.web.network.Request
+import fr.rob.web.network.RouteHandler
 import fr.rob.game.domain.opcode.OpcodeHandler as GameOpcodeHandler
 
 class OpcodeHandler(private val client: Client, private val opcodeHandler: GameOpcodeHandler) : RouteHandler() {
