+++ servers/web/src/main/kotlin/fr/rob/web/network/HttpServerHandler.kt
@@ -1,4 +1,4 @@
-package fr.rob.client.web.network
+package fr.rob.web.network
 
 import io.netty.buffer.ByteBuf
 import io.netty.buffer.Unpooled
