+++ servers/web/src/main/kotlin/fr/rob/web/network/Request.kt
@@ -1,4 +1,4 @@
-package fr.rob.client.web.network
+package fr.rob.web.network
 
 import io.netty.handler.codec.http.FullHttpRequest
 import io.netty.util.CharsetUtil
