+++ servers/web/src/main/kotlin/fr/rob/web/network/RouteHandler.kt
@@ -1,4 +1,4 @@
-package fr.rob.client.web.network
+package fr.rob.web.network
 
 
 abstract class RouteHandler {
